moved depth into its own file
