+++ src/render/pipeline/mod.rs
@@ -1,9 +1,11 @@
 mod buffer;
+mod depth;
 mod polygon;
 mod uniforms;
 mod vertex;
 
 use buffer::Buffer;
+use depth::DepthPipeline;
 use iced::{
     widget::shader::wgpu::{self, RenderPassDepthStencilAttachment},
     Rectangle, Size,
@@ -27,7 +29,7 @@ pub struct Pipeline {
     uniform_group: wgpu::BindGroup,
     depth_texture_size: Size<u32>,
     depth_view: wgpu::TextureView,
-    depth_pipeline: DepthPipeline,
+    depth_pipeline: depth::DepthPipeline,
     /// Actual number of vertices when drawn using Triangles
     vertex_count: u64,
     initialized: bool,
@@ -332,133 +334,7 @@ impl Pipeline {
             pass.set_bind_group(0, &self.uniform_group, &[]);
             pass.set_vertex_buffer(0, self.positions.raw.slice(..));
             pass.set_vertex_buffer(1, self.vertices.raw.slice(..));
-            //pass.set_vertex_buffer(2, self.transform.raw.slice(..));
             pass.draw(0..self.vertex_count as u32, 0..1);
         }
-    }
-}
-
-struct DepthPipeline {
-    bind_group_layout: wgpu::BindGroupLayout,
-    bind_group: wgpu::BindGroup,
-    sampler: wgpu::Sampler,
-    depth_view: wgpu::TextureView,
-}
-
-impl DepthPipeline {
-    pub fn new(
-        device: &wgpu::Device,
-        format: wgpu::TextureFormat,
-        depth_texture: wgpu::TextureView,
-    ) -> Self {
-        let sampler = device.create_sampler(&wgpu::SamplerDescriptor {
-            label: Some("cubes.depth_pipeline.sampler"),
-            ..Default::default()
-        });
-
-        let bind_group_layout = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-            label: Some("cubes.depth_pipeline.bind_group_layout"),
-            entries: &[
-                wgpu::BindGroupLayoutEntry {
-                    binding: 0,
-                    visibility: wgpu::ShaderStages::FRAGMENT,
-                    ty: wgpu::BindingType::Sampler(wgpu::SamplerBindingType::NonFiltering),
-                    count: None,
-                },
-                wgpu::BindGroupLayoutEntry {
-                    binding: 1,
-                    visibility: wgpu::ShaderStages::FRAGMENT,
-                    ty: wgpu::BindingType::Texture {
-                        sample_type: wgpu::TextureSampleType::Float { filterable: false },
-                        view_dimension: wgpu::TextureViewDimension::D2,
-                        multisampled: false,
-                    },
-                    count: None,
-                },
-            ],
-        });
-
-        let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-            label: Some("cubes.depth_pipeline.bind_group"),
-            layout: &bind_group_layout,
-            entries: &[
-                wgpu::BindGroupEntry {
-                    binding: 0,
-                    resource: wgpu::BindingResource::Sampler(&sampler),
-                },
-                wgpu::BindGroupEntry {
-                    binding: 1,
-                    resource: wgpu::BindingResource::TextureView(&depth_texture),
-                },
-            ],
-        });
-
-        let layout = device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
-            label: Some("cubes.depth_pipeline.layout"),
-            bind_group_layouts: &[&bind_group_layout],
-            push_constant_ranges: &[],
-        });
-
-        let shader = device.create_shader_module(wgpu::ShaderModuleDescriptor {
-            label: Some("cubes.depth_pipeline.shader"),
-            source: wgpu::ShaderSource::Wgsl(std::borrow::Cow::Borrowed(include_str!(
-                "../shaders/depth.wgsl"
-            ))),
-        });
-
-        let _pipeline = device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
-            label: Some("polyhedron.depth_pipeline.pipeline"),
-            layout: Some(&layout),
-            vertex: wgpu::VertexState {
-                module: &shader,
-                entry_point: "vs_main",
-                buffers: &[],
-            },
-            primitive: wgpu::PrimitiveState::default(),
-            depth_stencil: Some(wgpu::DepthStencilState {
-                format: wgpu::TextureFormat::Depth32Float,
-                depth_write_enabled: false,
-                depth_compare: wgpu::CompareFunction::Less,
-                stencil: wgpu::StencilState::default(),
-                bias: wgpu::DepthBiasState::default(),
-            }),
-            multisample: wgpu::MultisampleState::default(),
-            fragment: Some(wgpu::FragmentState {
-                module: &shader,
-                entry_point: "fs_main",
-                targets: &[Some(wgpu::ColorTargetState {
-                    format,
-                    blend: Some(wgpu::BlendState::REPLACE),
-                    write_mask: wgpu::ColorWrites::ALL,
-                })],
-            }),
-            multiview: None,
-        });
-
-        Self {
-            bind_group_layout,
-            bind_group,
-            sampler,
-            depth_view: depth_texture,
-        }
-    }
-
-    pub fn update(&mut self, device: &wgpu::Device, depth_texture: &wgpu::Texture) {
-        self.depth_view = depth_texture.create_view(&wgpu::TextureViewDescriptor::default());
-
-        self.bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-            label: Some("cubes.depth_pipeline.bind_group"),
-            layout: &self.bind_group_layout,
-            entries: &[
-                wgpu::BindGroupEntry {
-                    binding: 0,
-                    resource: wgpu::BindingResource::Sampler(&self.sampler),
-                },
-                wgpu::BindGroupEntry {
-                    binding: 1,
-                    resource: wgpu::BindingResource::TextureView(&self.depth_view),
-                },
-            ],
-        });
     }
 }
